Move libs to /lib/server to be explicit
