--- src/components/twitch/get-user-info.ts
@@ -1,7 +1,0 @@
-export const getUserInfo = async (oauthToken: string) => {
-    return fetch("https://id.twitch.tv/oauth2/userinfo", {
-        headers: {
-            Authorization: `Bearer ${oauthToken}`,
-        },
-    });
-};
