+++ src/lib/server/get-advanced-user-stats.ts
@@ -1,4 +1,5 @@
-export const advancedUserStats = async (user: string, timezone: string) => {
+"use server";
+export const getAdvancedUserStats = async (user: string, timezone: string) => {
     const res = await fetch(
         `https://dxg3hfz4b6ekynso45amgq2ife0tzhnl.lambda-url.eu-west-1.on.aws/${user}/${timezone}`,
     );
@@ -6,5 +7,3 @@ export const advancedUserStats = async (user: string, timezone: string) => {
 
     return json.result;
 };
-
-export default advancedUserStats;
