+++ src/lib/server/get-personal-best-runs.ts
@@ -1,4 +1,4 @@
-import { Run } from "../common/types";
+import { type Run } from "../../common/types";
 
 export const getPersonalBestRuns = async (): Promise<Run[]> => {
     const url = `${process.env.NEXT_PUBLIC_DATA_URL}/runs`;
