+++ src/lib/server/get-upload-key.ts
@@ -1,7 +1,6 @@
+"use server";
 export const getUploadKey = async (user: string) => {
     const url = `${process.env.NEXT_PUBLIC_DATA_URL}/users/uploadKey/${user}`;
 
     return (await fetch(url)).json();
 };
-
-export default getUploadKey;
